refactor: extracted per package models into models.go, refactored each package to follow the same handler, service, repository pattern
