+++ internal/auth/errors.go
