+++ internal/uploader/service.go
@@ -11,8 +11,8 @@ import (
 	"path/filepath"
 	"strings"
 	"time"
+	"volaticus-go/internal/common/models"
 	"volaticus-go/internal/config"
-	"volaticus-go/internal/models"
 )
 
 type Service struct {
