--- internal/uploader/wordlists.go
@@ -1,73 +1,0 @@
-package uploader
-
-// WordLists for generating GfyCat style URLs
-var (
-	adjectives = []string{
-		"adorable", "beautiful", "clever", "delightful", "elegant",
-		"fierce", "gentle", "happy", "intelligent", "jolly",
-		"kind", "lively", "magical", "noble", "peaceful",
-		"quick", "radiant", "silly", "talented", "unique",
-		"vibrant", "wise", "zealous", "brave", "calm",
-		"agile", "bright", "charming", "daring", "energetic",
-		"fearless", "graceful", "humble", "inventive", "joyful",
-		"keen", "luminous", "mighty", "neat", "optimistic",
-		"playful", "quirky", "resilient", "strong", "thoughtful",
-		"uplifting", "versatile", "whimsical", "youthful", "zesty",
-		"affectionate", "bold", "creative", "determined", "earnest",
-		"festive", "glorious", "hilarious", "inspiring", "judicious",
-		"knowledgeable", "loyal", "motivated", "nurturing", "outgoing",
-		"perceptive", "questioning", "resolute", "spontaneous", "tenacious",
-		"understanding", "valiant", "warm", "xenodochial", "zany",
-		"adaptable", "brilliant", "confident", "dazzling", "enthusiastic",
-		"friendly", "gentle", "honest", "imaginative", "jubilant",
-		"kindhearted", "lovable", "meticulous", "natural", "observant",
-		"persistent", "quickwitted", "reliable", "spirited", "thoughtprovoking",
-		"unassuming", "vivacious", "wonderful", "xenial", "zestful",
-	}
-
-	animals = []string{
-		"alpaca", "bear", "cat", "dolphin", "elephant",
-		"fox", "giraffe", "horse", "iguana", "jaguar",
-		"kangaroo", "lion", "monkey", "narwhal", "octopus",
-		"penguin", "quokka", "rabbit", "seal", "tiger",
-		"unicorn", "vulture", "whale", "xerus", "zebra",
-		"armadillo", "buffalo", "cheetah", "dog", "eel",
-		"flamingo", "goat", "hedgehog", "impala", "jellyfish",
-		"koala", "lemur", "moose", "newt", "ostrich",
-		"panda", "quail", "raccoon", "sheep", "toucan",
-		"urchin", "vole", "walrus", "xray fish", "yak",
-		"antelope", "beaver", "cougar", "deer", "emu",
-		"ferret", "gecko", "hamster", "ibis", "jackal",
-		"kudu", "llama", "marmot", "numbat", "oryx",
-		"platypus", "quokka", "reindeer", "salamander", "tapir",
-		"uakari", "vicuna", "wombat", "xenopus", "zebu",
-		"alligator", "bat", "crow", "dingo", "echidna",
-		"frog", "gull", "heron", "ibex", "jay",
-		"kingfisher", "lynx", "manatee", "nighthawk", "osprey",
-		"peacock", "quoll", "robin", "sparrow", "terrapin",
-		"urchin", "viper", "wolf", "xenops", "yellowtail",
-	}
-
-	colors = []string{
-		"amber", "blue", "crimson", "denim", "emerald",
-		"fuchsia", "gold", "hazel", "indigo", "jade",
-		"kotlin", "lavender", "maroon", "navy", "olive",
-		"purple", "quartz", "ruby", "silver", "teal",
-		"umber", "violet", "white", "xanthic", "yellow",
-		"aqua", "beige", "charcoal", "dusty rose", "eggplant",
-		"forest green", "grape", "honey", "ivory", "khaki",
-		"lime", "magenta", "neon green", "ocean", "peach",
-		"quince", "rose gold", "sapphire", "taupe", "umber",
-		"vermillion", "wine", "xenon blue", "blurple", "zinc",
-		"apricot", "bronze", "cobalt", "dandelion", "ebony",
-		"firebrick", "glacier", "harvest gold", "iris", "jasmine",
-		"kiwi", "lemon", "mint", "neon pink", "obsidian",
-		"pale blue", "quartz pink", "raspberry", "sand", "terracotta",
-		"ultramarine", "vanilla", "whisper white", "xanadu", "zircon",
-		"ash", "blush", "cerulean", "daffodil", "eucalyptus",
-		"flame", "granite", "heather", "ice", "jet",
-		"kelly green", "lilac", "moss", "nylon", "opal",
-		"pearl", "quasar yellow", "redwood", "slate", "topaz",
-		"umber", "velvet", "wheat", "xenic", "zenith blue",
-	}
-)
