+++ internal/user/errors.go
@@ -1,0 +1,11 @@
+package user
+
+import "errors"
+
+var (
+	ErrUserNotFound       = errors.New("user not found")
+	ErrEmailExists        = errors.New("email already exists")
+	ErrUsernameExists     = errors.New("username already exists")
+	ErrInvalidCredentials = errors.New("invalid credentials")
+	ErrInvalidInput       = errors.New("invalid input")
+)
