+++ internal/user/handlers.go
@@ -2,57 +2,60 @@ package user
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 	"net/http"
 )
 
 type Handler struct {
-	repo        UserRepository
+	service     *Service
 	authService AuthService
 }
 
 func NewHandler(repo UserRepository, authService AuthService) *Handler {
 	return &Handler{
-		repo:        repo,
+		service:     NewService(repo),
 		authService: authService,
 	}
 }
 
-type RegisterResponse struct {
-	User  *User  `json:"user"`
-	Token string `json:"token"`
+// CreateUserRequest represents the data needed to create a new user
+type CreateUserRequest struct {
+	Email    string `json:"email" validate:"required,email"`
+	Username string `json:"username" validate:"required,min=3,max=50"`
+	Password string `json:"password" validate:"required,min=8"`
+}
+
+// UpdateUserRequest represents the data that can be updated for a user
+type UpdateUserRequest struct {
+	Email    *string `json:"email" validate:"omitempty,email"`
+	Username *string `json:"username" validate:"omitempty,min=3,max=50"`
+	IsActive *bool   `json:"is_active"`
 }
 
 func (h *Handler) HandleRegister(w http.ResponseWriter, r *http.Request) {
 	var req CreateUserRequest
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		errorMsg := fmt.Sprintf("Invalid request body: %s", err.Error())
-		http.Error(w, errorMsg, http.StatusBadRequest)
-		return
-	}
-	log.Printf("%s", req)
-	// Basic validation
-	if req.Email == "" || req.Username == "" || req.Password == "" {
-		http.Error(w, "Email, username and password are required", http.StatusBadRequest)
+		http.Error(w, "Invalid request body", http.StatusBadRequest)
 		return
 	}
 
-	// Create user
-	user, err := h.repo.Create(&req)
+	// Create user using service
+	user, err := h.service.RegisterUser(&req)
 	if err != nil {
 		switch err {
 		case ErrEmailExists:
 			http.Error(w, "Email already exists", http.StatusConflict)
 		case ErrUsernameExists:
 			http.Error(w, "Username already exists", http.StatusConflict)
+		case ErrInvalidInput:
+			http.Error(w, "Invalid input", http.StatusBadRequest)
 		default:
-			log.Panic(err.Error())
+			log.Printf("Error registering user: %v", err)
 			http.Error(w, "Internal server error", http.StatusInternalServerError)
 		}
 		return
 	}
-	log.Printf("Created User")
+
 	// Generate JWT token
 	token, err := h.authService.GenerateToken(user)
 	if err != nil {
@@ -60,16 +63,15 @@ func (h *Handler) HandleRegister(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Redirect to /
 	// Set JWT cookie
 	http.SetCookie(w, &http.Cookie{
 		Name:     "jwt",
 		Value:    token,
 		Path:     "/",
 		HttpOnly: true,
-		Secure:   true, // Only send over HTTPS
+		Secure:   true,
 		SameSite: http.SameSiteStrictMode,
-		MaxAge:   3600 * 24, // 24 hours TODO: Implement token refresh
+		MaxAge:   3600 * 24,
 	})
-	w.Header().Set("HX-Redirect", "/")
+	//w.Header().Set("HX-Redirect", "/")
 }
