--- internal/user/user.go
@@ -1,48 +1,0 @@
-package user
-
-import (
-	"time"
-
-	"github.com/google/uuid"
-	"golang.org/x/crypto/bcrypt"
-)
-
-// User represents a user in the system
-type User struct {
-	ID           uuid.UUID `db:"id" json:"id"`
-	Email        string    `db:"email" json:"email"`
-	Username     string    `db:"username" json:"username"`
-	PasswordHash string    `db:"password_hash" json:"-"`
-	IsActive     bool      `db:"is_active" json:"is_active"`
-	CreatedAt    time.Time `db:"created_at" json:"created_at"`
-	UpdatedAt    time.Time `db:"updated_at" json:"updated_at"`
-}
-
-// UserRepository defines methods for user persistence
-type UserRepository interface {
-	Create(user *CreateUserRequest) (*User, error)
-	GetByID(id uuid.UUID) (*User, error)
-	GetByEmail(email string) (*User, error)
-	GetByUsername(username string) (*User, error)
-	Update(id uuid.UUID, updates *UpdateUserRequest) error
-	Delete(id uuid.UUID) error
-}
-
-// HashPassword creates a bcrypt hash from a password string
-func HashPassword(password string) (string, error) {
-	bytes, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return "", err
-	}
-	return string(bytes), nil
-}
-
-// CheckPassword checks if the provided password matches the hash
-func CheckPassword(password, hash string) bool {
-	err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
-	return err == nil
-}
-
-type AuthService interface {
-	GenerateToken(user *User) (string, error)
-}
